⚡️ Improve reactivity of devExtensionPath in commands rendering
When appConfig.devExtensionPath is updated in settings page, the main window listens to update on this field, and construct a new extension loader to load and re-render list. Add id field for each Extension loader (default to a uuid). Without setting a unique ID as key in vue v-for loop, new commands will not be rendered although they are already loaded.