⚡️ Pass metadata to iframe ext with URL
Previous design pass iframe ext identifier and manifest with pinia store, this works when extension is always in a single window, single instance. But  when we want to support  opening multiple ext in multiple window, the pinia store method won't work. If window is refreshed, it always loads the latest extension.